fix ts lint
